updated wallet UI and implementation
