+++ src/lib/utils/shorten.ts
@@ -1,7 +1,7 @@
 export const shorten = (address: string | undefined) => {
   if (!address) return "";
-  return `${address.substring(0, 6)}...${address.substring(
-    address.length - 4,
+  return `${address.substring(0, 8)}...${address.substring(
+    address.length - 5,
     address.length
   )}`;
 };
